Replace short for int in order to fit 6-digit verification token

+++ src/main/java/com/gregorioandrade/ds/accsync/data/DataConnector.java
@@ -4,7 +4,7 @@ import reactor.core.publisher.Mono;
 
 public interface DataConnector {
 
-    Mono<Void> createRequest(long discordId, short verificationToken);
+    Mono<Void> createRequest(long discordId, int verificationToken);
     Mono<Boolean> hasRequest(long discordId);
     Mono<Void> deleteRequest(long discordId);
 }
